Move virtual edge creation code out of QueryGraph
