+++ core/src/main/java/com/graphhopper/routing/QueryGraph.java
@@ -19,18 +19,18 @@ package com.graphhopper.routing;
 
 import com.carrotsearch.hppc.IntArrayList;
 import com.carrotsearch.hppc.IntObjectMap;
-import com.carrotsearch.hppc.predicates.IntObjectPredicate;
 import com.carrotsearch.hppc.procedures.IntProcedure;
 import com.graphhopper.coll.GHIntHashSet;
 import com.graphhopper.coll.GHIntObjectHashMap;
 import com.graphhopper.routing.util.AllEdgesIterator;
 import com.graphhopper.routing.util.EdgeFilter;
-import com.graphhopper.storage.*;
+import com.graphhopper.storage.Graph;
+import com.graphhopper.storage.GraphExtension;
+import com.graphhopper.storage.NodeAccess;
+import com.graphhopper.storage.TurnCostExtension;
 import com.graphhopper.storage.index.QueryResult;
 import com.graphhopper.util.*;
 import com.graphhopper.util.shapes.BBox;
-import com.graphhopper.util.shapes.GHPoint;
-import com.graphhopper.util.shapes.GHPoint3D;
 
 import java.util.*;
 
@@ -65,80 +65,7 @@ public class QueryGraph implements Graph {
      * Store lat,lon of virtual tower nodes.
      */
     private final PointList virtualNodes;
-    private final NodeAccess nodeAccess = new NodeAccess() {
-        @Override
-        public void ensureNode(int nodeId) {
-            mainNodeAccess.ensureNode(nodeId);
-        }
-
-        @Override
-        public boolean is3D() {
-            return mainNodeAccess.is3D();
-        }
-
-        @Override
-        public int getDimension() {
-            return mainNodeAccess.getDimension();
-        }
-
-        @Override
-        public double getLatitude(int nodeId) {
-            if (isVirtualNode(nodeId))
-                return virtualNodes.getLatitude(nodeId - mainNodes);
-            return mainNodeAccess.getLatitude(nodeId);
-        }
-
-        @Override
-        public double getLongitude(int nodeId) {
-            if (isVirtualNode(nodeId))
-                return virtualNodes.getLongitude(nodeId - mainNodes);
-            return mainNodeAccess.getLongitude(nodeId);
-        }
-
-        @Override
-        public double getElevation(int nodeId) {
-            if (isVirtualNode(nodeId))
-                return virtualNodes.getElevation(nodeId - mainNodes);
-            return mainNodeAccess.getElevation(nodeId);
-        }
-
-        @Override
-        public int getAdditionalNodeField(int nodeId) {
-            if (isVirtualNode(nodeId))
-                return 0;
-            return mainNodeAccess.getAdditionalNodeField(nodeId);
-        }
-
-        @Override
-        public void setNode(int nodeId, double lat, double lon) {
-            throw new UnsupportedOperationException("Not supported yet.");
-        }
-
-        @Override
-        public void setNode(int nodeId, double lat, double lon, double ele) {
-            throw new UnsupportedOperationException("Not supported yet.");
-        }
-
-        @Override
-        public void setAdditionalNodeField(int nodeId, int additionalValue) {
-            throw new UnsupportedOperationException("Not supported yet.");
-        }
-
-        @Override
-        public double getLat(int nodeId) {
-            return getLatitude(nodeId);
-        }
-
-        @Override
-        public double getLon(int nodeId) {
-            return getLongitude(nodeId);
-        }
-
-        @Override
-        public double getEle(int nodeId) {
-            return getElevation(nodeId);
-        }
-    };
+    private final NodeAccess nodeAccess;
 
     // Use LinkedHashSet for predictable iteration order.
     private final Set<VirtualEdgeIteratorState> unfavoredEdges = new LinkedHashSet<>(5);
@@ -156,14 +83,17 @@ public class QueryGraph implements Graph {
         mainNodes = graph.getNodes();
         mainEdges = graph.getEdges();
 
-        virtualEdges = new ArrayList<>(queryResults.size() * 2);
-        virtualNodes = new PointList(queryResults.size(), mainNodeAccess.is3D());
-        this.queryResults = new ArrayList<>(queryResults.size());
+        VirtualEdgeBuilder virtualEdgeBuilder = new VirtualEdgeBuilder(mainNodes, mainEdges, mainNodeAccess.is3D());
+        virtualEdgeBuilder.lookup(queryResults);
 
-        lookup(queryResults);
+        virtualEdges = virtualEdgeBuilder.getVirtualEdges();
+        virtualNodes = virtualEdgeBuilder.getVirtualNodes();
+        this.queryResults = virtualEdgeBuilder.getQueryResults();
+
+        nodeAccess = new ExtendedNodeAccess(mainNodeAccess, virtualNodes, mainNodes);
 
         if (mainGraph.getExtension() instanceof TurnCostExtension)
-            wrappedExtension = new QueryGraphTurnExt();
+            wrappedExtension = new QueryGraphTurnExt(mainGraph, this.queryResults);
         else
             wrappedExtension = mainGraph.getExtension();
 
@@ -191,151 +121,7 @@ public class QueryGraph implements Graph {
         virtualEdges = superQueryGraph.virtualEdges;
         virtualNodes = superQueryGraph.virtualNodes;
         queryResults = superQueryGraph.queryResults;
-    }
-
-    /**
-     * For all specified query results calculate snapped point and if necessary set closest node
-     * to a virtual one and reverse closest edge. Additionally the wayIndex can change if an edge is
-     * swapped.
-     *
-     * @see QueryGraph
-     */
-    private void lookup(List<QueryResult> resList) {
-        GHIntObjectHashMap<List<QueryResult>> edge2res = new GHIntObjectHashMap<>(resList.size());
-
-        // Phase 1
-        // calculate snapped point and swap direction of closest edge if necessary
-        for (QueryResult res : resList) {
-            // Do not create virtual node for a query result if it is directly on a tower node or not found
-            if (res.getSnappedPosition() == QueryResult.Position.TOWER)
-                continue;
-
-            EdgeIteratorState closestEdge = res.getClosestEdge();
-            if (closestEdge == null)
-                throw new IllegalStateException("Do not call QueryGraph.lookup with invalid QueryResult " + res);
-
-            int base = closestEdge.getBaseNode();
-
-            // Force the identical direction for all closest edges.
-            // It is important to sort multiple results for the same edge by its wayIndex
-            boolean doReverse = base > closestEdge.getAdjNode();
-            if (base == closestEdge.getAdjNode()) {
-                // check for special case #162 where adj == base and force direction via latitude comparison
-                PointList pl = closestEdge.fetchWayGeometry(0);
-                if (pl.size() > 1)
-                    doReverse = pl.getLatitude(0) > pl.getLatitude(pl.size() - 1);
-            }
-
-            if (doReverse) {
-                closestEdge = closestEdge.detach(true);
-                PointList fullPL = closestEdge.fetchWayGeometry(3);
-                res.setClosestEdge(closestEdge);
-                if (res.getSnappedPosition() == QueryResult.Position.PILLAR)
-                    // ON pillar node
-                    res.setWayIndex(fullPL.getSize() - res.getWayIndex() - 1);
-                else
-                    // for case "OFF pillar node"
-                    res.setWayIndex(fullPL.getSize() - res.getWayIndex() - 2);
-
-                if (res.getWayIndex() < 0)
-                    throw new IllegalStateException("Problem with wayIndex while reversing closest edge:" + closestEdge + ", " + res);
-            }
-
-            // find multiple results on same edge
-            int edgeId = closestEdge.getEdge();
-            List<QueryResult> list = edge2res.get(edgeId);
-            if (list == null) {
-                list = new ArrayList<>(5);
-                edge2res.put(edgeId, list);
-            }
-            list.add(res);
-        }
-
-        // Phase 2 - now it is clear which points cut one edge
-        // 1. create point lists
-        // 2. create virtual edges between virtual nodes and its neighbor (virtual or normal nodes)
-        edge2res.forEach(new IntObjectPredicate<List<QueryResult>>() {
-            @Override
-            public boolean apply(int edgeId, List<QueryResult> results) {
-                // we can expect at least one entry in the results
-                EdgeIteratorState closestEdge = results.get(0).getClosestEdge();
-                final PointList fullPL = closestEdge.fetchWayGeometry(3);
-                int baseNode = closestEdge.getBaseNode();
-                Collections.sort(results, new Comparator<QueryResult>() {
-                    @Override
-                    public int compare(QueryResult o1, QueryResult o2) {
-                        int diff = Integer.compare(o1.getWayIndex(), o2.getWayIndex());
-                        if (diff == 0) {
-                            return Double.compare(distanceOfSnappedPointToPillarNode(o1), distanceOfSnappedPointToPillarNode(o2));
-                        } else {
-                            return diff;
-                        }
-                    }
-                    private double distanceOfSnappedPointToPillarNode(QueryResult o) {
-                        GHPoint snappedPoint = o.getSnappedPoint();
-                        double fromLat = fullPL.getLatitude(o.getWayIndex());
-                        double fromLon = fullPL.getLongitude(o.getWayIndex());
-                        return Helper.DIST_PLANE.calcNormalizedDist(fromLat, fromLon, snappedPoint.lat, snappedPoint.lon);
-                    }
-                });
-
-                GHPoint3D prevPoint = fullPL.get(0);
-                int adjNode = closestEdge.getAdjNode();
-                int origEdgeKey = GHUtility.createEdgeKey(baseNode, adjNode, closestEdge.getEdge(), false);
-                int origRevEdgeKey = GHUtility.createEdgeKey(baseNode, adjNode, closestEdge.getEdge(), true);
-                int prevWayIndex = 1;
-                int prevNodeId = baseNode;
-                int virtNodeId = virtualNodes.getSize() + mainNodes;
-                boolean addedEdges = false;
-
-                // Create base and adjacent PointLists for all none-equal virtual nodes.
-                // We do so via inserting them at the correct position of fullPL and cutting the
-                // fullPL into the right pieces.
-                for (QueryResult res : results) {
-                    if (res.getClosestEdge().getBaseNode() != baseNode)
-                        throw new IllegalStateException("Base nodes have to be identical but were not: " + closestEdge + " vs " + res.getClosestEdge());
-
-                    GHPoint3D currSnapped = res.getSnappedPoint();
-
-                    // no new virtual nodes if exactly the same snapped point
-                    if (prevPoint.equals(currSnapped)) {
-                        res.setClosestNode(prevNodeId);
-                        continue;
-                    }
-
-                    queryResults.add(res);
-                    boolean isPillar = res.getSnappedPosition() == QueryResult.Position.PILLAR;
-                    createEdges(origEdgeKey, origRevEdgeKey,
-                            prevPoint, prevWayIndex, isPillar,
-                            res.getSnappedPoint(), res.getWayIndex(),
-                            fullPL, closestEdge, prevNodeId, virtNodeId);
-
-                    virtualNodes.add(currSnapped.lat, currSnapped.lon, currSnapped.ele);
-
-                    // add edges again to set adjacent edges for newVirtNodeId
-                    if (addedEdges) {
-                        virtualEdges.add(virtualEdges.get(virtualEdges.size() - 2));
-                        virtualEdges.add(virtualEdges.get(virtualEdges.size() - 2));
-                    }
-
-                    addedEdges = true;
-                    res.setClosestNode(virtNodeId);
-                    prevNodeId = virtNodeId;
-                    prevWayIndex = res.getWayIndex() + 1;
-                    prevPoint = currSnapped;
-                    virtNodeId++;
-                }
-
-                // two edges between last result and adjacent node are still missing if not all points skipped
-                if (addedEdges)
-                    createEdges(origEdgeKey, origRevEdgeKey,
-                            prevPoint, prevWayIndex, false,
-                            fullPL.get(fullPL.getSize() - 1), fullPL.getSize() - 2,
-                            fullPL, closestEdge, virtNodeId - 1, adjNode);
-
-                return true;
-            }
-        });
+        nodeAccess = superQueryGraph.nodeAccess;
     }
 
     @Override
@@ -370,39 +156,6 @@ public class QueryGraph implements Graph {
         this.useEdgeExplorerCache = useEECache;
         this.baseGraph.setUseEdgeExplorerCache(useEECache);
         return this;
-    }
-
-    private void createEdges(int origEdgeKey, int origRevEdgeKey,
-                             GHPoint3D prevSnapped, int prevWayIndex, boolean isPillar, GHPoint3D currSnapped, int wayIndex,
-                             PointList fullPL, EdgeIteratorState closestEdge,
-                             int prevNodeId, int nodeId) {
-        int max = wayIndex + 1;
-        PointList basePoints = new PointList(max - prevWayIndex + 1, mainNodeAccess.is3D());
-        basePoints.add(prevSnapped.lat, prevSnapped.lon, prevSnapped.ele);
-        for (int i = prevWayIndex; i < max; i++) {
-            basePoints.add(fullPL, i);
-        }
-        if (!isPillar) {
-            basePoints.add(currSnapped.lat, currSnapped.lon, currSnapped.ele);
-        }
-        // basePoints must have at least the size of 2 to make sure fetchWayGeometry(3) returns at least 2
-        assert basePoints.size() >= 2 : "basePoints must have at least two points";
-
-        PointList baseReversePoints = basePoints.clone(true);
-        double baseDistance = basePoints.calcDistance(Helper.DIST_PLANE);
-        int virtEdgeId = mainEdges + virtualEdges.size();
-
-        boolean reverse = closestEdge.get(EdgeIteratorState.REVERSE_STATE);
-        // edges between base and snapped point
-        VirtualEdgeIteratorState baseEdge = new VirtualEdgeIteratorState(origEdgeKey,
-                virtEdgeId, prevNodeId, nodeId, baseDistance, closestEdge.getFlags(), closestEdge.getName(), basePoints, reverse);
-        VirtualEdgeIteratorState baseReverseEdge = new VirtualEdgeIteratorState(origRevEdgeKey,
-                virtEdgeId, nodeId, prevNodeId, baseDistance, IntsRef.deepCopyOf(closestEdge.getFlags()), closestEdge.getName(), baseReversePoints, !reverse);
-
-        baseEdge.setReverseEdge(baseReverseEdge);
-        baseReverseEdge.setReverseEdge(baseEdge);
-        virtualEdges.add(baseEdge);
-        virtualEdges.add(baseReverseEdge);
     }
 
     /**
@@ -731,54 +484,5 @@ public class QueryGraph implements Graph {
 
     private UnsupportedOperationException exc() {
         return new UnsupportedOperationException("QueryGraph cannot be modified.");
-    }
-
-    class QueryGraphTurnExt extends TurnCostExtension {
-        private final TurnCostExtension mainTurnExtension;
-
-        public QueryGraphTurnExt() {
-            this.mainTurnExtension = (TurnCostExtension) mainGraph.getExtension();
-        }
-
-        @Override
-        public long getTurnCostFlags(int edgeFrom, int nodeVia, int edgeTo) {
-            if (isVirtualNode(nodeVia)) {
-                return 0;
-            } else if (isVirtualEdge(edgeFrom) || isVirtualEdge(edgeTo)) {
-                if (isVirtualEdge(edgeFrom)) {
-                    edgeFrom = getOriginalEdge(edgeFrom);
-                }
-                if (isVirtualEdge(edgeTo)) {
-                    edgeTo = getOriginalEdge(edgeTo);
-                }
-                return mainTurnExtension.getTurnCostFlags(edgeFrom, nodeVia, edgeTo);
-            } else {
-                return mainTurnExtension.getTurnCostFlags(edgeFrom, nodeVia, edgeTo);
-            }
-        }
-
-        @Override
-        public boolean isUTurn(int edgeFrom, int edgeTo) {
-            // detecting a u-turn from a virtual to a non-virtual edge requires looking at the original edge of the
-            // virtual edge. however when we are turning between virtual edges we need to compare the virtual edge ids
-            // see #1593
-            if (isVirtualEdge(edgeFrom) && !isVirtualEdge(edgeTo)) {
-                edgeFrom = getOriginalEdge(edgeFrom);
-            } else if (!isVirtualEdge(edgeFrom) && isVirtualEdge(edgeTo)) {
-                edgeTo = getOriginalEdge(edgeTo);
-            }
-            return mainTurnExtension.isUTurn(edgeFrom, edgeTo);
-        }
-
-        @Override
-        public boolean isUTurnAllowed(int node) {
-            // do not allow u-turns at virtual nodes, otherwise the route depends on whether or not there are virtual
-            // via nodes, see #1672
-            return !isVirtualNode(node);
-        }
-
-        private int getOriginalEdge(int edgeFrom) {
-            return queryResults.get((edgeFrom - mainEdges) / 4).getClosestEdge().getEdge();
-        }
     }
 }
